requirements.txt
python-dotenv>=1.0.0
websockets>=11.0.3 
mcp>=1.8.1
pydantic>=2.11.4
mcp-proxy>=0.8.2
lark-oapi>=1.4.8
paho-mqtt>=2.1.0
schedule>=1.2.2